api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.2",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "langserve>=0.3.1",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.0.2",
    "uvicorn>=0.35.0",
]
